lsmachineinfo: add minimal tool to get machineinfo
This is way simpler than runningthe full cadvisor,
even though, admittedly, running cadivsor is not that big of a deal
in the first place.

However, this utility, besides being minimal, emits the machineinfo JSON
and it emits it filtered by default - tries hard to skip any
machine-identifiable part, which is unnecessary for `cpumgrx` anyway.

Signed-off-by: Francesco Romani <fromani@redhat.com>